refactor: add logger util
