+++ src/utils/helpers/exportFile.ts
@@ -8,9 +8,7 @@ type FileProps = {
 
 export function getExportFileBlob(props: FileProps) {
   const { columns, data, fileType } = props;
-  console.log(columns);
-  console.log(data);
-  console.log(fileType);
+
   if (fileType === "csv") {
     const headerNames = columns.map((col: any) => col.exportValue);
     const csvString = Papa.unparse({ fields: headerNames, data });
