+++ src/utils/helpers/logger.ts
@@ -1,0 +1,7 @@
+const log = (...data: any) => {
+  if (process.env.REACT_APP_DEVELOPMENT) {
+    console.log(data);
+  }
+};
+
+export default log;
